DolphinServer/requirements.txt
fastapi
uvicorn[standard]
pycryptodome
bs4
aiohttp
aiomysql
aiosmtplib
ics
email